Custom 404 template for widgets
We probably don't want the default 404 template showing up in
broken embed widgets, so use a custom template,
``storybase_story/widget_404.html`` for widget 404 errors.

This functionality is implemented via a mixin class,
``Custom404Mixin`` that short-circuits the default handling
of ``Http404`` exceptions.  This should be used sparingly.

Also moved ``VersionTemplateMixin`` to the
``storybase.views.generic`` module.

Addresses #741
