+++ apps/storybase_story/templates/storybase_story/widget_404.html
@@ -1,0 +1,8 @@
+{# TODO: Put something in here! #}
+<html>
+    <head>
+        <title></title>
+    </head>
+    <body>
+    </body>
+</html>
